rewrite snackplayer with remark plugin
